refactor: split handler dispatch map into feature groups
