+++ assistant/src/daemon/handlers/publish.ts
@@ -10,7 +10,7 @@ import type {
   PublishPageRequest,
   UnpublishPageRequest,
 } from '../ipc-protocol.js';
-import { log, requestSecretStandalone, type HandlerContext } from './shared.js';
+import { log, requestSecretStandalone, type HandlerContext, type DispatchMap } from './shared.js';
 
 export async function handlePublishPage(
   msg: PublishPageRequest,
@@ -180,3 +180,8 @@ export async function handleUnpublishPage(
     });
   }
 }
+
+export const publishHandlers: Partial<DispatchMap> = {
+  publish_page: handlePublishPage,
+  unpublish_page: handleUnpublishPage,
+};
